mv python script
